+++ Classes/Growler.h
@@ -9,7 +9,8 @@
 
 #define GROWLER [[NSApp delegate] growler]
 
-@interface Growler : NSObject<GrowlApplicationBridgeDelegate>
+@interface Growler : NSObject<GrowlApplicationBridgeDelegate,
+                              NSUserNotificationCenterDelegate>
 
 - (void) growl:(Song*)song withImage:(NSData*)image isNew:(BOOL) n;